+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - File tab pack forget when there aren't any tabs
 - New theme system with json files
 - When suggestion is mapped and hit ENTER will autocomplete code
+- Ocean theme
 
 ### Fixed: